+++ tests/test_global_opts.rs
@@ -125,6 +125,26 @@ fn test_resolve_workspace_directory() {
     Working copy : 230dd059e1b0 (no description set)
     The working copy is clean
     "###);
+
+    // Explicit subdirectory path
+    let stderr = test_env.jj_cmd_failure(&subdir, &["status", "-R", "."]);
+    insta::assert_snapshot!(stderr, @r###"
+    Error: There is no jj repo in "."
+    "###);
+
+    // Valid explicit path
+    let stdout = test_env.jj_cmd_success(&subdir, &["status", "-R", "../.."]);
+    insta::assert_snapshot!(stdout, @r###"
+    Parent commit: 000000000000 (no description set)
+    Working copy : 230dd059e1b0 (no description set)
+    The working copy is clean
+    "###);
+
+    // "../../..".ancestors() contains "../..", but it should never be looked up.
+    let stderr = test_env.jj_cmd_failure(&subdir, &["status", "-R", "../../.."]);
+    insta::assert_snapshot!(stderr, @r###"
+    Error: There is no jj repo in "../../.."
+    "###);
 }
 
 #[test]